rag-chat/requirements.txt
numpy
fastapi
uvicorn
python-dotenv

faiss-cpu

langchain
langgraph 
langchain-core 
langchain-cohere 
langchain-community 